backend/requirements.txt
fastapi
uvicorn[standard]
python-socketio
SQLAlchemy
# For AI and Vector DB
torch
transformers
Pillow
chromadb
